requirements.txt
Django>=4.2
django-widget-tweaks==1.4.8
Pillow
whitenoise
gunicorn
dj-database-url
psycopg2-binary
<<<<<<< HEAD
python-dotenv
=======
python-dotenv
>>>>>>> 34b84ca (auto create superuser)
